Disable unused npm dependents request and remove uses of npm-stats.
Since the registry no longer supports querying dependents against
CouchDB and we have been showing a warning on attempts to do this
we can remove this request and the dependency we used for it.

Note that the codepath is kept in place in case another source of
data for this is found but the fetchNpmDependents() method becomes
hard-coded to return no result.